feat(matrix)!: refactor constructor, add full() class method
The `Matrix` constructor now accepts a `list[list[float]]` for initialization.
The previous constructor arguments (`rows`, `cols`, `fill_value`) have been moved to a new class method `Matrix.full()`.
This also updates the `total_cost` field in `Assignment` to `f64` for improved precision.

BREAKING CHANGE: The `zignal.Matrix` constructor signature has changed. It now takes `data: list[list[float]]`. Use `zignal.Matrix.full(rows, cols, fill_value)` for the old behavior.

+++ src/optimization.zig
@@ -24,7 +24,7 @@ pub const Assignment = struct {
     /// null means row i has no assignment
     assignments: []?usize,
     /// Total cost of the assignment
-    total_cost: f32,
+    total_cost: f64,
     /// Allocator used for assignments array
     allocator: Allocator,
 
@@ -317,7 +317,7 @@ pub fn solveAssignmentProblem(
     }
 
     // Calculate total cost and prepare result
-    var total_cost: f32 = 0;
+    var total_cost: f64 = 0;
     var result_assignments = try allocator.alloc(?usize, n_rows);
     for (0..n_rows) |i| {
         if (row_assignment[i]) |col| {
@@ -325,7 +325,7 @@ pub fn solveAssignmentProblem(
                 result_assignments[i] = col;
                 // Use original cost matrix values (not the multiplied work matrix)
                 const cost_val = cost_matrix.at(i, col).*;
-                total_cost += as(f32, cost_val);
+                total_cost += as(f64, cost_val);
             } else {
                 result_assignments[i] = null;
             }
